+++ src/Product/ConfigurableProductInterface.php
@@ -6,10 +6,49 @@ namespace Lsv\Datapump\Product;
 
 interface ConfigurableProductInterface
 {
+    public function setConfigurableAttributeKeys(array $keys);
+
+    public function getConfigurableAttributeKeys(): array;
+
+    public function addProduct(AbstractProduct $product);
+
     /**
+     * @param AbstractProduct[] $products
+     */
+    public function setProducts(array $products);
+
+    /**
+     * @return AbstractProduct[]
+     */
+    public function getProducts(): array;
+
+    public function countProducts(): int;
+
+    /**
+     * @deprecated
+     *
+     * @param AbstractProduct $product
+     */
+    public function addSimpleProduct(AbstractProduct $product);
+
+    /**
+     * @deprecated
+     *
      * @return AbstractProduct[]
      */
     public function getSimpleProducts(): array;
 
+    /**
+     * @deprecated
+     *
+     * @param AbstractProduct[] $products
+     */
+    public function setSimpleProducts(array $products);
+
+    /**
+     * @deprecated
+     *
+     * @return int
+     */
     public function countSimpleProducts(): int;
 }